ai-news-summarizer/requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
flask-bcrypt==1.0.1
python-dateutil==2.8.2